+++ crates/graphql/src/schema/objects/nft.rs
@@ -6,8 +6,8 @@ use indexer_core::{
         sql_query,
         sql_types::Text,
         tables::{
-            attributes, auction_houses, bid_receipts, collection_mints, current_metadata_owners,
-            listing_receipts, listings, metadata_collection_keys, metadata_jsons, metadatas,
+            attributes, auction_houses, bid_receipts, listing_receipts, listings,
+            metadata_collection_keys, metadata_jsons, metadatas,
         },
     },
     pubkeys,
@@ -516,7 +516,7 @@ If no value is provided, it will return XSmall")))]
             .map_err(Into::into)
     }
 
-    pub async fn collection(&self, ctx: &AppContext) -> FieldResult<Option<Collection>> {
+    pub async fn collection(&self, ctx: &AppContext) -> FieldResult<Option<CollectionNFT>> {
         ctx.nft_collection_loader
             .load(self.address.clone().into())
             .await
@@ -545,16 +545,16 @@ If no value is provided, it will return XSmall")))]
 }
 
 #[derive(Debug, Clone)]
-pub struct Collection(pub Nft);
+pub struct CollectionNFT(pub Nft);
 
-impl From<Nft> for Collection {
+impl From<Nft> for CollectionNFT {
     fn from(nft: Nft) -> Self {
         Self(nft)
     }
 }
 
 #[graphql_object(Context = AppContext)]
-impl Collection {
+impl CollectionNFT {
     fn nft(&self) -> &Nft {
         &self.0
     }
@@ -853,7 +853,7 @@ impl Collection {
     }
 
     #[graphql(deprecated = "use `nft { collection }`")]
-    pub async fn collection(&self, ctx: &AppContext) -> FieldResult<Option<Collection>> {
+    pub async fn collection(&self, ctx: &AppContext) -> FieldResult<Option<CollectionNFT>> {
         ctx.nft_collection_loader
             .load(self.0.address.clone().into())
             .await
@@ -882,7 +882,7 @@ impl Collection {
     }
 }
 
-impl TryFrom<models::Nft> for Collection {
+impl TryFrom<models::Nft> for CollectionNFT {
     type Error = <Nft as TryFrom<models::Nft>>::Error;
 
     fn try_from(value: models::Nft) -> Result<Self, Self::Error> {
@@ -1285,211 +1285,5 @@ impl NftsStats {
             .context("failed to load listed nfts count")?;
 
         Ok(count.try_into()?)
-    }
-}
-
-pub type CollectionId = String;
-
-#[derive(Debug, Clone)]
-pub struct Coll {
-    pub id: CollectionId,
-    pub image: String,
-    pub name: String,
-    pub description: String,
-    pub twitter_url: Option<String>,
-    pub discord_url: Option<String>,
-    pub website_url: Option<String>,
-    pub magic_eden_id: Option<String>,
-    pub verified_collection_address: Option<String>,
-    pub pieces: i32,
-    pub verified: bool,
-    pub go_live_at: DateTime<Utc>,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-}
-
-impl<'a> TryFrom<models::Collection<'a>> for Coll {
-    type Error = std::num::TryFromIntError;
-
-    fn try_from(
-        models::Collection {
-            id,
-            image,
-            name,
-            description,
-            twitter_url,
-            discord_url,
-            website_url,
-            magic_eden_id,
-            verified_collection_address,
-            pieces,
-            verified,
-            go_live_at,
-            created_at,
-            updated_at,
-        }: models::Collection,
-    ) -> Result<Self, Self::Error> {
-        Ok(Self {
-            id: id.to_string(),
-            image: image.to_string(),
-            name: name.to_string(),
-            description: description.to_string(),
-            twitter_url: twitter_url.map(Into::into),
-            discord_url: discord_url.map(Into::into),
-            website_url: website_url.map(Into::into),
-            magic_eden_id: magic_eden_id.map(Into::into),
-            verified_collection_address: verified_collection_address.map(Into::into),
-            pieces: pieces.try_into().unwrap_or_default(),
-            verified,
-            go_live_at: DateTime::from_utc(go_live_at, Utc),
-            created_at: DateTime::from_utc(created_at, Utc),
-            updated_at: DateTime::from_utc(updated_at, Utc),
-        })
-    }
-}
-
-#[graphql_object(Context = AppContext)]
-impl Coll {
-    pub fn id(&self) -> &CollectionId {
-        &self.id
-    }
-
-    pub fn image(&self) -> &str {
-        &self.image
-    }
-
-    pub fn name(&self) -> &str {
-        &self.name
-    }
-
-    pub fn description(&self) -> &str {
-        &self.description
-    }
-
-    pub fn twitter_url(&self) -> Option<&str> {
-        self.twitter_url.as_deref()
-    }
-
-    pub fn discord_url(&self) -> Option<&str> {
-        self.discord_url.as_deref()
-    }
-
-    pub fn website_url(&self) -> Option<&str> {
-        self.website_url.as_deref()
-    }
-
-    pub fn magic_eden_id(&self) -> Option<&str> {
-        self.magic_eden_id.as_deref()
-    }
-
-    pub fn verified_collection_address(&self) -> Option<&str> {
-        self.verified_collection_address.as_deref()
-    }
-
-    pub fn pieces(&self) -> i32 {
-        self.pieces
-    }
-
-    pub fn verified(&self) -> bool {
-        self.verified
-    }
-
-    pub fn go_live_at(&self) -> DateTime<Utc> {
-        self.go_live_at
-    }
-
-    pub fn created_at(&self) -> DateTime<Utc> {
-        self.created_at
-    }
-
-    pub fn updated_at(&self) -> DateTime<Utc> {
-        self.updated_at
-    }
-
-    // nfts data loader
-
-    pub async fn nfts(&self, ctx: &AppContext, limit: i32, offset: i32) -> FieldResult<Vec<Nft>> {
-        let conn = ctx.shared.db.get()?;
-
-        let nfts: Vec<models::Nft> = metadatas::table
-            .inner_join(
-                metadata_jsons::table.on(metadata_jsons::metadata_address.eq(metadatas::address)),
-            )
-            .inner_join(
-                current_metadata_owners::table
-                    .on(current_metadata_owners::mint_address.eq(metadatas::mint_address)),
-            )
-            .inner_join(
-                collection_mints::table.on(metadatas::mint_address.eq(collection_mints::mint)),
-            )
-            .filter(collection_mints::collection_id.eq(self.id.clone()))
-            .select(queries::metadatas::NFT_COLUMNS)
-            .offset(offset.into())
-            .limit(limit.into())
-            .load(&conn)
-            .context("Failed to load NFTs")?;
-
-        nfts.into_iter()
-            .map(TryInto::try_into)
-            .collect::<Result<_, _>>()
-            .map_err(Into::into)
-    }
-
-    pub fn attribute_groups(&self, context: &AppContext) -> FieldResult<Vec<AttributeGroup>> {
-        let conn = context.shared.db.get()?;
-
-        let metadata_attributes: Vec<models::MetadataAttribute> = attributes::table
-            .inner_join(metadatas::table.on(attributes::metadata_address.eq(metadatas::address)))
-            .inner_join(
-                collection_mints::table.on(metadatas::mint_address.eq(collection_mints::mint)),
-            )
-            .filter(collection_mints::collection_id.eq(&self.id))
-            .select(attributes::all_columns)
-            .load(&conn)
-            .context("Failed to load metadata attributes")?;
-
-        services::attributes::group(metadata_attributes)
-    }
-
-    #[graphql(description = "Count of NFTs in the collection.")]
-    async fn nft_count(&self, context: &AppContext) -> FieldResult<Option<scalars::I64>> {
-        Ok(context
-            .mr_collection_nft_count_loader
-            .load(self.id.clone().into())
-            .await?
-            .map(|dataloaders::collection::CollectionNftCount(nft_count)| nft_count))
-    }
-
-    #[graphql(
-        description = "Count of wallets that currently hold at least one NFT from the collection."
-    )]
-    pub async fn holder_count(&self, ctx: &AppContext) -> FieldResult<Option<scalars::I64>> {
-        Ok(ctx
-            .mr_collection_holders_count_loader
-            .load(self.id.clone().into())
-            .await?
-            .map(|dataloaders::collection::CollectionHoldersCount(nft_count)| nft_count))
-    }
-
-    pub async fn activities(
-        &self,
-        ctx: &AppContext,
-        event_types: Option<Vec<String>>,
-        limit: i32,
-        offset: i32,
-    ) -> FieldResult<Vec<NftActivity>> {
-        let conn = ctx.shared.db.get()?;
-        let rows = queries::collections::mr_collection_activities(
-            &conn,
-            &self.id,
-            event_types,
-            limit,
-            offset,
-        )?;
-
-        rows.into_iter()
-            .map(TryInto::try_into)
-            .collect::<Result<_, _>>()
-            .map_err(Into::into)
     }
 }
